pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
openai = "^0.27.8"
fastapi = {extras = ["all"], version = "^0.100.1"}
aerich = "^0.7.2"
aiohttp = "^3.8.5"
redis = "^4.6.0"
aiogram = "^3.0b8"
pydantic-settings = "^2.0.2"
uvicorn = {extras = ["standart"], version = "^0.23.2"}



[tool.aerich]
tortoise_orm = "src.settings.DATABASE_CONFIG"
location = "src/db/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
